requirements.txt
langchain~=0.0.152
setuptools~=67.7.2
python-dotenv~=1.0.0